update layout structure
