+++ types/dxf/config.d.ts
@@ -1,0 +1,3 @@
+export interface config {
+  verbose: boolean;
+}
