+++ types/dxf/denormalise.d.ts
@@ -1,0 +1,4 @@
+import * as Information from './Information';
+import * as Entities from './handlers/entities';
+
+export function denormalise(parsed: Information.FileInfo): Entities.Entity[];
