+++ types/dxf/groupEntitiesByLayer.d.ts
@@ -1,0 +1,3 @@
+import * as Entities from './handlers/entities';
+
+export function groupEntitiesByLayer(entities: Entities.Entity[]): Entities.LayerGroupedEntities;
