+++ types/dxf/handlers/entity/Circle.d.ts
@@ -1,6 +1,6 @@
-import { Point3D } from "../Common";
+import { Point3D } from "../../Common";
 import { CommonEntityData } from "./Common";
 
-export type VertexEntityData = {
-  bulge?: number;
+export type CircleEntityData = {
+  r?: number;
 } & Partial<Point3D> & Partial<CommonEntityData>;
