+++ types/dxf/handlers/entity/Solid.d.ts
@@ -1,4 +1,4 @@
-import { Point3D } from '../Common';
+import { Point3D } from "../../Common";
 import { CommonEntityData } from './Common';
 
 export type SolidEntityData = {
