+++ types/dxf/handlers/entity/Vertex.d.ts
@@ -1,6 +1,6 @@
-import { Point3D } from "../Common";
+import { Point3D } from "../../Common";
 import { CommonEntityData } from "./Common";
 
-export type CircleEntityData = {
-  r?: number;
+export type VertexEntityData = {
+  bulge?: number;
 } & Partial<Point3D> & Partial<CommonEntityData>;
