+++ types/dxf/handlers/tables.d.ts
@@ -1,5 +1,5 @@
-import * as Layers from "./Layers";
-import * as Styles from "./Styles";
+import * as Layers from "../Layers";
+import * as Styles from "../Styles";
 
 export interface DXFTable {
   layers: Layers.Layer[];
