+++ types/dxf/index.d.ts
@@ -5,26 +5,40 @@
 // Minimum TypeScript Version: 4.1
 
 import Helper from './Helper';
-import * as Blocks from './Blocks';
-import * as Common from './Common';
-import * as Entities from './Entities';
-import * as Header from './Header';
+import * as Blocks from './handlers/blocks';
+import { ColorNumbers as colors } from './Common';
+import * as Entities from './handlers/entities';
+import * as Header from './handlers/header';
 import * as Information from './Information';
 import * as Layers from './Layers';
 import * as Styles from './Styles';
-import * as Tables from './Tables';
+import * as Tables from './handlers/tables';
 import * as Utils from './Utils';
+import { config } from './config';
+import { parseString } from './parseString';
+import { denormalise } from './denormalise';
+import { groupEntitiesByLayer } from './groupEntitiesByLayer';
+import { toPolylines } from './toPolylines';
+import { toSVG } from './toSVG';
 
-export interface config {
-  verbose: boolean;
-}
+export {
+  config,
+  parseString,
+  denormalise,
+  groupEntitiesByLayer,
+  toPolylines,
+  toSVG,
+  colors,
+  Helper
+};
 
-export function parseString(string: string): Information.FileInfo;
-export function denormalise(parsed: Information.FileInfo): Entities.Entity[];
-export function groupEntitiesByLayer(entities: Entities.Entity[]): Entities.LayerGroupedEntities;
-export function toPolylines(parsed: Information.FileInfo): Information.Polyline[];
-export function toSVG(parsed: Information.FileInfo): Information.SVG;
-export const colors: Common.ColorNumbers;
-export default Helper;
-
-export { Blocks, Common, Entities, Header, Information, Layers, Styles, Tables, Utils };
+export {
+  Blocks,
+  Entities,
+  Header,
+  Information,
+  Layers,
+  Styles,
+  Tables,
+  Utils
+};
