+++ types/dxf/parseString.d.ts
@@ -1,0 +1,3 @@
+import * as Information from './Information';
+
+export function parseString(string: string): Information.FileInfo;
