+++ types/dxf/toPolylines.d.ts
@@ -1,0 +1,3 @@
+import * as Information from './Information';
+
+export function toPolylines(parsed: Information.FileInfo): Information.Polyline[];
