+++ types/dxf/toSVG.d.ts
@@ -1,0 +1,3 @@
+import * as Information from './Information';
+
+export function toSVG(parsed: Information.FileInfo): Information.SVG;
