+++ types/dxf/tsconfig.json
@@ -18,33 +18,40 @@
     "forceConsistentCasingInFileNames": true
   },
   "files": [
-    "dxf-tests.ts",
-    "index.d.ts",
-    "Blocks.d.ts",
     "Common.d.ts",
-    "Entities.d.ts",
-    "Header.d.ts",
+    "config.d.ts",
+    "denormalise.d.ts",
+    "dxf-tests.ts",
+    "groupEntitiesByLayer.d.ts",
+    "Common.d.ts",
     "Helper.d.ts",
+    "index.d.ts",
     "Information.d.ts",
     "Layers.d.ts",
+    "parseString.d.ts",
     "Styles.d.ts",
-    "Tables.d.ts",
+    "toPolylines.d.ts",
+    "toSVG.d.ts",
     "Utils.d.ts",
-    "./entities/Arc.d.ts",
-    "./entities/Circle.d.ts",
-    "./entities/Common.d.ts",
-    "./entities/Dimension.d.ts",
-    "./entities/Ellipse.d.ts",
-    "./entities/Insert.d.ts",
-    "./entities/Line.d.ts",
-    "./entities/LWPolyline.d.ts",
-    "./entities/MText.d.ts",
-    "./entities/Point.d.ts",
-    "./entities/Polyline.d.ts",
-    "./entities/Solid.d.ts",
-    "./entities/Spline.d.ts",
-    "./entities/Text.d.ts",
-    "./entities/ThreeDFace.d.ts",
-    "./entities/Vertex.d.ts"
+    "./handlers/blocks.d.ts",
+    "./handlers/entities.d.ts",
+    "./handlers/header.d.ts",
+    "./handlers/tables.d.ts",
+    "./handlers/entity/Arc.d.ts",
+    "./handlers/entity/Circle.d.ts",
+    "./handlers/entity/Common.d.ts",
+    "./handlers/entity/Dimension.d.ts",
+    "./handlers/entity/Ellipse.d.ts",
+    "./handlers/entity/Insert.d.ts",
+    "./handlers/entity/Line.d.ts",
+    "./handlers/entity/LWPolyline.d.ts",
+    "./handlers/entity/MText.d.ts",
+    "./handlers/entity/Point.d.ts",
+    "./handlers/entity/Polyline.d.ts",
+    "./handlers/entity/Solid.d.ts",
+    "./handlers/entity/Spline.d.ts",
+    "./handlers/entity/Text.d.ts",
+    "./handlers/entity/ThreeDFace.d.ts",
+    "./handlers/entity/Vertex.d.ts"
   ]
 }
